GVT-1517: Separate sorting logic into its own file and clean up publication-table
